Move to proof directory
