+++ chain/client/proof/abi.go
@@ -1,4 +1,4 @@
-package rpc
+package proof
 
 // TODO: It will be changed after update new bridge contract
 var relayFormat = []byte(`
